+++ CHANGELOG.md
@@ -138,6 +138,7 @@
 - Cookie should use httpOnly and Secure (#355)
 - crash in import-mbox when list-id is missing and --lid is not provided (#358)
 - archiver traverses multi-part message parts twice (#359)
+- Add unsubscribe button (#362)
 
 
 ## CHANGES in 0.9b: